refactor: BoardView를 통해서 책임 이전 및 함수 중복 제거
