+++ README.md
@@ -52,7 +52,7 @@
 ## 3차 피드백 사항
 - [x] Closable 객체를 통해 db close
 - [x] 사용하지 않는 함수및 인자 제거
-- [ ] forEachIndexed 로 어떠한 행동을 할지만 함수로 받으서 중복 코드 해결
+- [x] BoardView를 통해서 책임 이전 및 함수 중복 제거
 - [x] 게임 도메인 생성시 보드화면 초기화
 - [ ] 끝날 경우 초기화 하는 로직 이동
 - [ ] Game 테스트 구현
