+++ app/src/main/java/woowacourse/omok/domain/event/GameEvent.kt
@@ -1,7 +1,10 @@
 package woowacourse.omok.domain.event
 
+import woowacourse.omok.domain.Game
 import woowacourse.omok.domain.model.stone.Stones
 
 interface GameEvent {
-    fun initBoard(stones: Stones)
+    fun initBoard(omokGame: Game)
+
+    fun showInitStones(stones: Stones)
 }
