+++ app/src/main/java/woowacourse/ui/mapper/PositionMapper.kt
@@ -1,7 +1,8 @@
-package woowacourse.ui
+package woowacourse.ui.mapper
 
 import woowacourse.omok.domain.model.Board
 import woowacourse.omok.domain.model.position.Position
+import woowacourse.ui.model.PositionUiModel
 
 fun PositionUiModel.toPosition(board: Board) =
     Position(
