+++ app/src/main/java/woowacourse/ui/model/PositionUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.ui
+package woowacourse.ui.model
 
 data class PositionUiModel(
     val column: Int,
